Move job execution out of Backstage.Job
The `Backstage.Job` module is only concerned with the representation of
a job  and interfacing with the database so it doesn't make sense to
keep this small bit here.

`Backstage.Consumer` is already handling much of the logic around
running job and their underlying task so it feels natural to have that
piece of logic there.